Update titles and status of existing AEPs
The titles of all existing AEPs are updated to be prefixed with the
explicit AEP number. The automated documentation will use the titles for
the table of contents, and having the AEP number in front in a
consistent manner will make it easier to search and browse.

In addition, the status of some AEPs that have now been implemented have
been updated accordingly.