add 'draft' status
Especially with larger projects it can take a significant amount of time
(easily a couple of months) between first discussions on an enhancement
proposal and its actual implementation/release in aiida-core.  With the
current approach, we have typically seen some initial discussion in a
pull request, which then died at some point.

Since there is no "checkpointing", when looking back at the PR after a
long time, it can be unclear what was actually agreed upon and which
questions were left open. Agreeing to merge the PR in a draft state
provides such a checkpoint (besides making the draft more visible
in the AEP repository).

We also remove the 'submitted' status, since it was used only during
the pull request stage and always needed to be changed before merging.